requirements.txt
feedparser>=6.0
requests>=2.31
langchain>=0.3
langchain-google-genai>=4.2
google-genai>=1.6
python-dotenv>=1.0
